Norm, norm and norms!
Missed beats and extra beats have been traditionally
normalised with sentivitity and accuracy. So there is
no point of inventing the wheel again and they are
between 0..1.

Jitter needs to refer to time to be hardware independent.
When averaging over different detectors their jitter
varies wildly so not a good measure.

However  Cassirame et al. 2019 state that 4ms is a
dealbreaker. Still, that just 4ms will lead to a
relaive jitter of one so we need the mapping function.

So the JMX score is:
score = mapping(jitter / 4ms) * accuracy